its too early 4 this
realized i wrote episode list code inside of the episode card component switched code
